+++ apps/fabric-website/src/pages/Controls/PopupMenuPage/PopupMenuPage.tsx
@@ -5,7 +5,7 @@ import { PopupMenuPageProps } from './PopupMenuPage.doc';
 import { Platforms } from '../../../interfaces/Platforms';
 
 const baseUrl =
-  'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/apps/fabric-website/src/pages/Controls/PopupMenuPage/';
+  'https://github.com/microsoft/fluentui/tree/master/apps/fabric-website/src/pages/Controls/PopupMenuPage/';
 
 export const PopupMenuPage: React.FunctionComponent<IControlsPageProps> = props => {
   const { platform } = props;