+++ apps/fabric-website/src/pages/Styles/Colors/MessagingPage.doc.ts
@@ -1,8 +1,7 @@
 import { TFabricPlatformPageProps } from '../../../interfaces/Platforms';
 
 const title = 'Messaging';
-const componentUrl =
-  'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/apps/fabric-website/src/pages/Styles/Colors';
+const componentUrl = 'https://github.com/microsoft/fluentui/tree/master/apps/fabric-website/src/pages/Styles/Colors';
 
 export const ColorsMessagingPageProps: TFabricPlatformPageProps = {
   web: {