+++ packages/charting/src/components/StackedBarChart/StackedBarChart.types.ts
@@ -93,6 +93,11 @@ export interface IStackedBarChartProps {
   legendsOverflowProps?: Partial<IOverflowSetProps>;
 
   /**
+   * text for overflow legends string
+   */
+  legendsOverflowText?: string;
+
+  /**
    * focus zone props in hover card for legends
    */
   focusZonePropsForLegendsInHoverCard?: IFocusZoneProps;