+++ packages/fluentui/react-northstar/src/themes/teams/componentVariables.ts
@@ -103,3 +103,10 @@ export { default as CarouselNavigationItem } from './components/Carousel/carouse
 export { default as Table } from './components/Table/tableVariables';
 export { default as TableRow } from './components/Table/tableRowVariables';
 export { default as TableCell } from './components/Table/tableCellVariables';
+
+export { default as Card } from './components/Card/cardVariables';
+export { default as CardHeader } from './components/Card/cardHeaderVariables';
+export { default as CardBody } from './components/Card/cardBodyVariables';
+export { default as CardFooter } from './components/Card/cardFooterVariables';
+export { default as CardPreview } from './components/Card/cardPreviewVariables';
+export { default as CardTopControls } from './components/Card/cardTopControlsVariables';