+++ packages/office-ui-fabric-react/src/components/Breadcrumb/Breadcrumb.doc.tsx
@@ -13,7 +13,7 @@ export const BreadcrumbPageProps: IDocPageProps = {
   title: 'Breadcrumb',
   componentName: 'Breadcrumb',
   componentUrl:
-    'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/packages/office-ui-fabric-react/src/components/Breadcrumb',
+    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/Breadcrumb',
   examples: [
     {
       title: 'Breadcrumb rendering options',