+++ packages/react/src/Dialog.ts
@@ -1,2 +1,1 @@
 export * from 'office-ui-fabric-react/lib/Dialog';
-export { Dialog as default } from 'office-ui-fabric-react/lib/Dialog';