+++ packages/react/src/Modal.ts
@@ -1,2 +1,1 @@
 export * from 'office-ui-fabric-react/lib/Modal';
-export { Modal as default } from 'office-ui-fabric-react/lib/Modal';